pyproject.toml
[tool.poetry]
name = "skillarena-chat"
version = "0.1.0"
description = ""
authors = ["ChiragAgg5k <110609663+ChiragAgg5k@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
jinja2 = "^3.1.4"
logging = "^0.4.9.6"
fastapi-limiter = "^0.1.6"
motor = "^3.5.1"
asyncio = "^3.4.3"
python-jose = "^3.3.0"
boto3 = "^1.34.144"
botocore = "^1.34.144"
pillow = "^10.4.0"
exponent-server-sdk = "^2.1.0"
pypdf2 = "^3.0.1"
pycryptodome = "^3.20.0"

[tool.poetry.scripts]
start = "uvicorn skillarena_chat.main:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
